Medisas-Twitter
Use Twitter's sample streaming API to show the top 10 retweeted tweets
in a rolling window of time, where the window's start is n minutes ago
(where n is defined by the user) and the window's end is the current
time.

Output should include the tweet text and number of times retweeted in
the current rolling window.